# openapi.json
{
  "components": {
    "schemas": {
...
      "CreateWorkspaceRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "path"
        ],
        "type": "object"
      },
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "reason"
        ],
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/Error"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "StatusResponse": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "type": "object"
      },
...
      "UpdateWorkspaceRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Workspace": {
        "additionalProperties": false,
        "properties": {
          "client_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "path",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "consoleSession": {
        "in": "cookie",
        "name": "wingman_session",
        "type": "apiKey"
      },
      "rootBearer": {
        "description": "Daemon root credential",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Wingman API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/workspaces": {
      "get": {
        "operationId": "listWorkspaces",
        "parameters": [
          {
            "description": "Client identity for resource attribution and scoping",
            "in": "header",
            "name": "X-Wingman-Client",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Workspace"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request failed"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "consoleSession": []
          }
        ],
        "summary": "List Workspaces"
      },
      "post": {
        "operationId": "createWorkspace",
        "parameters": [
          {
            "description": "Client identity for resource attribution and scoping",
            "in": "header",
            "name": "X-Wingman-Client",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request failed"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "consoleSession": []
          }
        ],
        "summary": "Create a Workspace"
      }
    },
    "/workspaces/{id}": {
      "delete": {
        "operationId": "deleteWorkspace",
        "parameters": [
          {
            "description": "Client identity for resource attribution and scoping",
            "in": "header",
            "name": "X-Wingman-Client",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request failed"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "consoleSession": []
          }
        ],
        "summary": "Delete a Workspace"
      },
      "get": {
        "operationId": "getWorkspace",
        "parameters": [
          {
            "description": "Client identity for resource attribution and scoping",
            "in": "header",
            "name": "X-Wingman-Client",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request failed"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "consoleSession": []
          }
        ],
        "summary": "Get a Workspace"
      },
      "put": {
        "operationId": "updateWorkspace",
        "parameters": [
          {
            "description": "Client identity for resource attribution and scoping",
            "in": "header",
            "name": "X-Wingman-Client",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Request failed"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "consoleSession": []
          }
        ],
        "summary": "Update a Workspace"
      }
    },
...
  }
}